feat(tool): port pure-data primitives & param helpers to dasclaw_tool (#670)
Verbatim move of the tool-layer pure-data primitives and parameter
helpers from ironclaw to the shared dasclaw_tool crate so any dasclaw
host (desktop, CLI, headless, admin backend) can describe and dispatch
tool invocations without depending on the ironclaw desktop crate.

Moved (verbatim, ADR-129 §1.3):
- ApprovalRequirement, ApprovalContext
- RiskLevel, ToolDomain, ToolRateLimitConfig
- ToolOutput, ToolSchema, ToolDiscoverySummary
- require_str, require_param, redact_params, validate_tool_schema

Not in this PR (deferred to follow-up):
- Tool trait — couples to crate::context::JobContext and
  crate::tools::wasm::WebhookCapability; needs a separate decoupling ADR
- LspQueryTool — depends on Tool trait being available in dasclaw_tool

Compatibility:
- ironclaw side re-exports all moved items via 'pub use dasclaw_tool::{..}'
- Public API surface unchanged: existing import paths still compile
- dasclaw_tool bumped to edition 2024 to preserve let-chain idioms

Cross-cuts: 类A (no new .ironclaw / IRONCLAW_BASE_DIR literals)

Refs #670
